Map (#23)
* Map in Maze1.unity

Scene has a map using the bricks prefab and grass prefab. Should be able to put this directly into unity if the prefabs already exist in your assets folder. I will be adding another map to this but feel free to test traps and stuff in the meantime!

* Add files via upload

* Delete Maze1.unity

* Tile made map

used TileMaps to make maps, both maps are in the same scene. ALL tiles made from asset package from Klyoman included.

Co-authored-by: dcwell <cwell.de@gmail.com>